skip `dataFlowAnalyzer.checkType` no TypeResolutionInterceptors are installed
I believe the lack of `dataFlowAnalyzer.checkType` in this locoation is a bug, but to
minimize potential to introduce inadvertent bugs, hide this behind an EP.
